control/voice_control.py
import os
import google.generativeai as genai
import chainlit as cl
from .automation_switch import is_system_active, is_safety_checks_enabled

# Configure Gemini Flash API
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
model = genai.GenerativeModel('gemini-pro')

def process_voice(audio_path):
    if not is_system_active() or not is_safety_checks_enabled():
        return None

    # Implement speech-to-text logic here
    text = "Placeholder for speech-to-text conversion"

    # Generate content using Gemini
    response = model.generate_content(text)
    return response.text

def start_voice_control():
    if is_system_active() and is_safety_checks_enabled():
        # Implement logic to start voice control
        pass

def stop_voice_control():
    if is_system_active() and is_safety_checks_enabled():
        # Implement logic to stop voice control
        pass

def update_voice_control_ui():
    with cl.sidebar():
        cl.title("Voice Control")
        cl.button("Start Voice Control", on_click=lambda: start_voice_control())
        cl.button("Stop Voice Control", on_click=lambda: stop_voice_control())
