control/requirements.txt
chainlit==0.7.700
google-generativeai==0.3.1
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
mkdocs==1.5.3 